include,Additions: Windows 10 tweaks.
svn:sync-xref-src-repo-rev: r96474

+++ src/VBox/Additions/WINNT/Graphics/Video/mp/xpdm/VBoxMPInternal.cpp
@@ -1,4 +1,4 @@
-/* $Id: VBoxMPInternal.cpp 49750 2013-12-03 12:31:51Z vitali.pelenjow@oracle.com $ */
+/* $Id: VBoxMPInternal.cpp 53008 2014-10-09 11:34:57Z vitali.pelenjow@oracle.com $ */
 
 /** @file
  * VBox XPDM Miniport internal functions
@@ -426,7 +426,7 @@ static int VBoxVbvaCreateChannelContexts(PVBOXMP_COMMON pCommon, VBVA_CHANNELCON
     VBVA_CHANNELCONTEXTS *pContext = (VBVA_CHANNELCONTEXTS*) VBoxMPMemAllocDriver(pCommon, size);
     if (pContext)
     {
-        VideoPortZeroMemory(pContext, size);
+        VideoPortZeroMemory(pContext, (ULONG)size);
         pContext->cContexts = cDisplays;
         pContext->pCommon = pCommon;
         *ppContext = pContext;
